requirements (1).txt
fastapi
uvicorn
pymongo
passlib
python-dotenv